Merge of hdtCat

+++ hdt-java-core/src/main/java/org/rdfhdt/hdt/cache/DictionaryCacheLRI.java
@@ -39,30 +39,30 @@ import java.util.concurrent.ConcurrentHashMap;
  */
 public class DictionaryCacheLRI<T> implements DictionaryCache<T> {
 
-	private Map<Integer, T> cache;
-	private int [] arr;
+	private Map<Long, T> cache;
+	private long [] arr;
 	private int ptr=0;
 	private final int size;
 	
 	public DictionaryCacheLRI(int size) {
 		this.size = size;
-		arr = new int[size];
-		cache = new ConcurrentHashMap<Integer, T>(size);
+		arr = new long[size];
+		cache = new ConcurrentHashMap<Long, T>(size);
 	}
 		
 	/* (non-Javadoc)
 	 * @see hdt.jena.DictionaryNodeCache#getNode(int)
 	 */
 	@Override
-	public T get(int id) {
+	public T get(long id) {
 		return cache.get(id);
 	}
 
 	/* (non-Javadoc)
-	 * @see hdt.jena.DictionaryNodeCache#setNode(int, com.hp.hpl.jena.graph.Node)
+	 * @see hdt.jena.DictionaryNodeCache#setNode(long, T)
 	 */
 	@Override
-	public void put(int id, T node) {
+	public void put(long id, T node) {
 		cache.put(id, node);
 		if(cache.size()>size) {
 			cache.remove(arr[ptr]);
