+++ hdt-java-core/src/main/java/org/rdfhdt/hdt/iterator/DictionaryTranslateIterator.java
@@ -48,7 +48,7 @@ public class DictionaryTranslateIterator implements IteratorTripleString {
 
 	CharSequence s, p, o;
 	
-	int lastSid, lastPid, lastOid;
+	long lastSid, lastPid, lastOid;
 	CharSequence lastSstr, lastPstr, lastOstr;
 	
 	/**
